third_party/WebKit/LayoutTests/accessibility/crash-determining-aria-role-when-label-present.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body id="body">

<label><q role=x><input id="test">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests a crashing scenario where an element with a role attribute is a child of a label that also has a corresponding control.");

    if (window.accessibilityController) {
          document.getElementById("test").focus();
         
          // This line should not crash.
          var input = accessibilityController.focusedElement;

          shouldBe("input.childrenCount", "1");
    }

</script>

</body>
</html>
